requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
google-auth==2.48.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
gspread-formatting==1.2.1
idna==3.11
numpy==2.4.2
oauthlib==3.3.1
pandas==3.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.17.0
six==1.17.0
urllib3==2.6.3
